Update User API

+++ user-service/src/main/kotlin/com/hulkdx/findprofessional/repository/UserRepository.kt
@@ -3,6 +3,6 @@ package com.hulkdx.findprofessional.repository
 import com.hulkdx.findprofessional.model.User
 import org.springframework.data.repository.kotlin.CoroutineCrudRepository
 
-interface UserRepository : CoroutineCrudRepository<User, Int> {
+interface UserRepository : CoroutineCrudRepository<User, Long> {
     suspend fun findByEmail(email: String): User?
 }
